Re-did the whole project
I decided to discard almost everything except the stuff I already
refactored and re-built the project. I added a package structure,
superclasses for Bullets and GameCharacters and just in general tried to
tidy up as much as possible.
Currently both the Enemy (circle only) and the Player are working as
they should, except that they don't take damage.